Stop commiting if no unrelated content was removed
The script removes files outside a 'subdir' folder by a sequence of
commands `git rm -rf` and `git checkout` and commit the changes.
If there was no files outside the subdirectory there was nothing to
commit and the script failed.

Additionialy the commit adds a few useful debug logs and unifies
the string quotation marks.